# tlias-web-management/assert/Web开发-AI.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Web开发-AI",
    "description": "新版本的Web开发的接口文档",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/clazzs/{id}": {
      "get": {
        "summary": "根据ID查询",
        "x-apifox-folder": "班级管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "根据ID查询班级信息",
        "operationId": "getInfoUsingGET",
        "tags": [
          "班级管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "example": "1",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-apifox-refs": {
                    "01H7YTX04J7GX93MGMS37EFYGC": {
                      "$ref": "#/components/schemas/Result"
                    }
                  },
                  "x-apifox-orders": [
                    "01H7YTX04J7GX93MGMS37EFYGC"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer",
                      "title": "状态码",
                      "description": "1: 成功 , 0: 失败",
                      "format": "int32"
                    },
                    "msg": {
                      "type": "string",
                      "title": "错误提示信息",
                      "nullable": true
                    },
                    "data": {
                      "title": "响应数据",
                      "anyOf": [
                        {
                          "$ref": "#/components/schemas/Clazz"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "x-apifox-ignore-properties": [
                    "code",
                    "msg",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3161925/apis/api-103410787-run"
      },
      "delete": {
        "summary": "删除班级",
        "x-apifox-folder": "班级管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "根据ID删除班级信息",
        "operationId": "deleteUsingDELETE",
        "tags": [
          "班级管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "example": "1",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3161925/apis/api-103410788-run"
      }
    },
...
    "/emps/{id}": {
      "get": {
        "summary": "根据ID查询",
        "x-apifox-folder": "员工管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "根据ID查询员工",
        "operationId": "getInfoUsingGET_1",
        "tags": [
          "员工管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "example": "1",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-apifox-refs": {},
                  "x-apifox-orders": [
                    "code",
                    "msg",
                    "data"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer",
                      "title": "状态码",
                      "description": "1: 成功 , 0: 失败",
                      "format": "int32",
                      "additionalProperties": false
                    },
                    "msg": {
                      "type": "string",
                      "title": "错误提示信息",
                      "additionalProperties": false,
                      "nullable": true
                    },
                    "data": {
                      "title": "响应数据",
                      "type": "object",
                      "x-apifox-refs": {},
                      "x-apifox-orders": [
                        "id",
                        "username",
                        "password",
                        "name",
                        "gender",
                        "job",
                        "salary",
                        "image",
                        "entryDate",
                        "deptId",
                        "createTime",
                        "updateTime",
                        "exprList"
                      ],
                      "properties": {
                        "id": {
                          "type": "integer",
                          "title": "ID",
                          "format": "int32",
                          "description": "ID",
                          "additionalProperties": false,
                          "nullable": true
                        },
                        "username": {
                          "type": "string",
                          "title": "用户名",
                          "description": "用户名",
                          "additionalProperties": false,
                          "nullable": true
                        },
                        "password": {
                          "type": "string",
                          "title": "密码",
                          "description": "密码",
                          "additionalProperties": false,
                          "nullable": true
                        },
                        "name": {
                          "type": "string",
                          "title": "姓名",
                          "description": "姓名",
                          "additionalProperties": false,
                          "nullable": true
                        },
                        "gender": {
                          "type": "integer",
                          "title": "性别",
                          "description": "性别, 1: 男 , 2: 女",
                          "format": "int32",
                          "additionalProperties": false,
                          "nullable": true
                        },
                        "job": {
                          "type": "integer",
                          "title": "职位",
                          "description": "职位, 1 班主任,2 讲师, 3 学工主管, 4 教研主管, 5 咨询师",
                          "format": "int32",
                          "additionalProperties": false,
                          "nullable": true
                        },
                        "salary": {
                          "type": "integer",
                          "format": "int32",
                          "title": "薪资",
                          "description": "薪资",
                          "additionalProperties": false
                        },
                        "image": {
                          "type": "string",
                          "title": "头像",
                          "description": "头像",
                          "additionalProperties": false,
                          "nullable": true
                        },
                        "entryDate": {
                          "type": "string",
                          "title": "入职日期",
                          "format": "date",
                          "description": "入职日期",
                          "additionalProperties": false,
                          "nullable": true
                        },
                        "deptId": {
                          "type": "integer",
                          "title": "部门ID",
                          "format": "int32",
                          "description": "部门ID",
                          "additionalProperties": false,
                          "nullable": true
                        },
                        "createTime": {
                          "type": "string",
                          "title": "创建时间",
                          "format": "date-time",
                          "description": "创建时间",
                          "additionalProperties": false,
                          "nullable": true
                        },
                        "updateTime": {
                          "type": "string",
                          "title": "修改时间",
                          "format": "date-time",
                          "description": "修改时间",
                          "additionalProperties": false,
                          "nullable": true
                        },
                        "exprList": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/EmpExpr"
                          },
                          "title": "工作经历",
                          "description": "工作经历"
                        }
                      },
                      "required": [
                        "salary"
                      ],
                      "x-apifox-ignore-properties": [],
                      "nullable": true
                    }
                  },
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3161925/apis/api-103410799-run"
      }
    },
...
    "/students/{id}": {
      "get": {
        "summary": "根据ID查询",
        "x-apifox-folder": "学生管理",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "根据ID查询",
        "operationId": "getInfoUsingGET_2",
        "tags": [
          "学生管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "example": "1",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "x-apifox-refs": {
                    "01H7Z1NS5AJB5W2JG6FFQ3NGYE": {
                      "$ref": "#/components/schemas/Result"
                    }
                  },
                  "x-apifox-orders": [
                    "01H7Z1NS5AJB5W2JG6FFQ3NGYE"
                  ],
                  "properties": {
                    "code": {
                      "type": "integer",
                      "title": "状态码",
                      "description": "1: 成功 , 0: 失败",
                      "format": "int32"
                    },
                    "msg": {
                      "type": "string",
                      "title": "错误提示信息",
                      "nullable": true
                    },
                    "data": {
                      "title": "响应数据",
                      "anyOf": [
                        {
                          "$ref": "#/components/schemas/Student"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "x-apifox-ignore-properties": [
                    "code",
                    "msg",
                    "data"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apifox-orders": [],
                  "x-apifox-ignore-properties": []
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3161925/apis/api-103410809-run"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmpExpr": {
        "$comment": "JSON Schema for emp_expr table",
        "$id": "emp_expr",
        "title": "",
        "description": "员工工作经历表",
        "type": "object",
        "properties": {
          "id": {
            "$comment": "primary key",
            "description": "ID, 主键",
            "type": "integer",
            "minimum": 1,
            "maximum": 4294967296,
            "title": "ID, 主键"
          },
          "begin": {
            "description": "开始时间",
            "type": "string",
            "format": "date",
            "title": "开始时间"
          },
          "end": {
            "description": "结束时间",
            "type": "string",
            "format": "date",
            "title": "结束时间"
          },
          "company": {
            "description": "公司名字",
            "type": "string",
            "maxLength": 50,
            "title": "公司名字"
          },
          "job": {
            "description": "职位",
            "type": "string",
            "maxLength": 50,
            "title": "职位"
          },
          "empId": {
            "description": "员工ID",
            "type": "integer",
            "minimum": 0,
            "maximum": 4294967296,
            "title": "员工ID"
          }
        },
        "x-apifox-orders": [
          "id",
          "begin",
          "end",
          "company",
          "job",
          "empId"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "Schemas"
      },
...
      "Student": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "title": "ID",
            "additionalProperties": false
          },
          "name": {
            "type": "string",
            "title": "姓名",
            "additionalProperties": false
          },
          "no": {
            "type": "string",
            "title": "学号",
            "additionalProperties": false
          },
          "gender": {
            "type": "integer",
            "format": "int32",
            "title": "性别",
            "description": "1: 男 , 2: 女",
            "additionalProperties": false
          },
          "phone": {
            "type": "string",
            "title": "手机号",
            "additionalProperties": false
          },
          "idCard": {
            "type": "string",
            "title": "身份证号",
            "additionalProperties": false
          },
          "isCollege": {
            "type": "integer",
            "title": "是否是院校学生",
            "description": "1: 是 , 0: 否",
            "format": "int32"
          },
          "address": {
            "type": "string",
            "title": "联系地址",
            "additionalProperties": false
          },
          "degree": {
            "type": "integer",
            "format": "int32",
            "title": "最高学历",
            "description": "1:初中, 2:高中, 3:大专, 4:本科, 5:硕士, 6:博士",
            "additionalProperties": false
          },
          "graduationDate": {
            "type": "string",
            "format": "date-time",
            "title": "毕业时间",
            "additionalProperties": false
          },
          "clazzId": {
            "type": "integer",
            "format": "int32",
            "title": "班级ID",
            "additionalProperties": false
          },
          "violationCount": {
            "type": "integer",
            "format": "int32",
            "title": "违纪次数",
            "additionalProperties": false
          },
          "violationScore": {
            "type": "integer",
            "format": "int32",
            "title": "违纪扣分",
            "additionalProperties": false
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "title": "创建时间",
            "additionalProperties": false
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "title": "修改时间",
            "additionalProperties": false
          }
        },
        "title": "Student",
        "x-apifox-orders": [
          "id",
          "name",
          "no",
          "gender",
          "phone",
          "idCard",
          "isCollege",
          "address",
          "degree",
          "graduationDate",
          "clazzId",
          "violationCount",
          "violationScore",
          "createTime",
          "updateTime"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "Schemas"
      },
      "Clazz": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "title": "ID",
            "description": "ID"
          },
          "name": {
            "type": "string",
            "title": "班级名称",
            "description": "班级"
          },
          "room": {
            "type": "string",
            "title": "班级教室",
            "description": "班级教室"
          },
          "beginDate": {
            "type": "string",
            "format": "date",
            "title": "开课时间",
            "description": "开课时间"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "title": "结课时间",
            "description": "结课时间"
          },
          "masterId": {
            "type": "integer",
            "format": "int32",
            "title": "班主任id",
            "description": "班主任id, 关联员工表id"
          },
          "subject": {
            "type": "integer",
            "format": "int32",
            "title": "学科",
            "description": "1:java, 2:前端, 3:大数据, 4:Python, 5:Go, 6:嵌入式"
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "title": "创建时间",
            "description": "创建时间"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "title": "修改时间",
            "description": "修改时间"
          }
        },
        "title": "Clazz",
        "x-apifox-orders": [
          "id",
          "name",
          "room",
          "beginDate",
          "endDate",
          "masterId",
          "subject",
          "createTime",
          "updateTime"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "Schemas"
      },
...
      "Result": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "title": "状态码",
            "description": "1: 成功 , 0: 失败",
            "format": "int32"
          },
          "msg": {
            "type": "string",
            "title": "错误提示信息",
            "nullable": true
          },
          "data": {
            "type": "object",
            "properties": {},
            "title": "响应数据",
            "x-apifox-orders": [],
            "x-apifox-ignore-properties": [],
            "nullable": true
          }
        },
        "title": "Result",
        "x-apifox-orders": [
          "code",
          "msg",
          "data"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "Schemas"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "开发环境"
    }
  ]
}